Move a module and rename an interface
Fixes a cyclical dependency and prevents a name conflict in the
grpc_server module
